added some checks
Made it so that 'set' doesn't print out a bunch of crap, and allowed support for $MODULE_DIR to be a symbolic link. 

The loop that set up the MODULES array would never add any of the modules into the array, so the check for number of modules in the $MODULE array would always fail. It was fixed when the while loop was changed to a for loop. go figure.

Also, fixed a few spelling/grammar mistakes.